refactor(claude_statusline): extract models and segments
* Renames ClaudePayload to StatusLineStdIn
* Extracts pydantic models into models.py
* Re-adds all documented fields from the claude statusline docs
* Extracts rendering components into segments.py
* Refactors rendering layout loop into render.py
* Creates tests package and moves tests inside
