+++ src/python/claude_statusline/claude_statusline/segments/workspace.py
@@ -1,0 +1,26 @@
+from pathlib import Path
+
+from claude_statusline.segments.constants import BLUE, ICON_DIR, RESET, Segment
+
+
+def shorten_path(path: Path) -> str:
+    """Shortens the path for display (e.g. ~/projects/foo -> .../projects/foo)."""
+    try:
+        rel_path = path.relative_to(Path.home())
+        parts = list(rel_path.parts)
+        if len(parts) > 3:
+            return f".../{parts[-2]}/{parts[-1]}"
+        if str(rel_path) == ".":
+            return "~"
+        return f"~/{rel_path}"
+    except ValueError:
+        parts = list(path.parts)
+        if len(parts) > 3:
+            return f".../{parts[-2]}/{parts[-1]}"
+        return str(path)
+
+
+def format_directory(cwd: Path) -> Segment | None:
+    display_path = shorten_path(cwd)
+    cwd_link = f"\033]8;;file://{cwd}\033\\{display_path}\033]8;;\033\\"
+    return Segment(f"{BLUE}{ICON_DIR} {cwd_link}{RESET}")
